Provide BaseGoalSpider
