+++ epl_scrapy/epl_scrapy/spiders/base.py
@@ -1,7 +1,7 @@
 import datetime as dt
 
 import scrapy
-from epl_scrapy.items import MatchItem
+from epl_scrapy.items import GoalItem, MatchItem
 
 
 class BaseMatchSpider(scrapy.Spider):
@@ -81,3 +81,46 @@ class BaseMatchSpider(scrapy.Spider):
 
     def _parse_referee(self, response):
         return response.xpath("string(//div[@class='referee'])").get().strip()
+
+
+class BaseGoalSpider(scrapy.Spider):
+    name = "epl_goals"
+
+    def start_requests(self):
+        for start_url in self.start_urls:
+            yield scrapy.Request(
+                url=start_url,
+                callback=self.parse_goals,
+                cb_kwargs={"match_id": start_url.split("/")[-1]}
+            )
+    
+    def parse_goals(self, response, match_id):
+        xpath_to_events = "//div[@class='matchEvents matchEventsContainer']/*/div[@class='event']"
+        
+        # To count number of events
+        events = response.xpath(xpath_to_events)
+        for idx in range(1, len(events)+1):
+            xpath_to_each_event = f"({xpath_to_events})[{idx}]"
+            player = response.xpath(
+                f"{xpath_to_each_event}/*/text()"
+            ).get()
+            minutes = response.xpath(
+                f"{xpath_to_each_event}/a/following-sibling::text()[1]"
+            ).get().strip().split(", ")
+            # It was either goal or own goal or red card or second yellow card.
+            label = response.xpath(
+                f"{xpath_to_each_event}/div/div/span/text()"
+            ).get()
+            # Determine it was scored by home or away team.
+            side = response.xpath(
+                f"{xpath_to_each_event}/parent::node()/@class"
+            ).get()
+            for minute in minutes:
+                goal = GoalItem(
+                    match_id=match_id,
+                    player=player,
+                    minute=minute,
+                    label=label,
+                    side=side,
+                )
+                yield goal
